// src/models/DTO/match/websocket/WebsocketErrorEnum.ts
export enum WebsocketErrorEnum {
  RETRY_GENERAL = "RETRY_GENERAL",
  CLOSE_CONNECTION_GENERAL = "CLOSE_CONNECTION_GENERAL",
  CLOSE_CONNECTION_ALREADY_IN_MATCH = "CLOSE_CONNECTION_ALREADY_IN_MATCH",
  CLOSE_CONNECTION_NO_ACTIVE_MATCH = "CLOSE_CONNECTION_NO_ACTIVE_MATCH",
  CLOSE_CONNECTION_ALREADY_SUBSCRIBED = "CLOSE_CONNECTION_ALREADY_SUBSCRIBED",
}